+++ corpus/c/macro_include/case-12/src/main.c
@@ -1,0 +1,10 @@
+#include "wrapper.h"
+
+typedef struct {
+    int value;
+} Counter;
+
+int main(void) {
+    Counter counter = {1};
+    return OUTER_ACCESS(counter);
+}