+++ corpus/c/type/case-11/src/main.c
@@ -1,0 +1,11 @@
+void takes_int_ptr(int *value);
+
+struct Box {
+    int value;
+};
+
+int main(void) {
+    struct Box box = {1};
+    takes_int_ptr(&box);
+    return 0;
+}