+++ corpus/c/type/case-12/src/main.c
@@ -1,0 +1,9 @@
+struct Box {
+    int value;
+};
+
+int main(void) {
+    struct Box box = {1};
+    int value = box;
+    return value;
+}