+++ corpus/cpp/overload/case-07/src/main.cpp
@@ -1,0 +1,6 @@
+void takes_short(short);
+void takes_long(long);
+
+int main() {
+    takes_short("x");
+}