+++ corpus/cpp/template/case-14/src/main.cpp
@@ -1,0 +1,9 @@
+template <typename T>
+struct Pair {
+    Pair(T, T) {}
+};
+
+int main() {
+    Pair value(1, "x");
+    (void)value;
+}